add bare minimum tests
